+++ Bundles/Search/src/Spryker/Client/Search/Plugin/Elasticsearch/ResultFormatter/PaginatedResultFormatterPlugin.php
@@ -40,7 +40,7 @@ class PaginatedResultFormatterPlugin extends AbstractElasticsearchResultFormatte
             ->getPaginationConfigBuilder();
 
         $itemsPerPage = $paginationConfig->getCurrentItemsPerPage($requestParameters);
-        $maxPage = ceil($searchResult->getTotalHits() / $itemsPerPage);
+        $maxPage = (int)ceil($searchResult->getTotalHits() / $itemsPerPage);
         $currentPage = min($paginationConfig->getCurrentPage($requestParameters), $maxPage);
 
         $paginationSearchResultTransfer = new PaginationSearchResultTransfer();
